cambios antes de hacer el merge 14 04 2022 con la rama daniela

+++ resorte.cpp
@@ -2,7 +2,7 @@
 
 
 
-double resorte::getVelocidad() const
+float resorte::getVelocidad() const
 {
     return velocidad;
 }
